skills/preparation_recommendation/resources/requirements.txt
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
